% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_plot.R
\name{dualplot}
\alias{dualplot}
\title{#' FUNCOES PHDRISk}
\usage{
dualplot(
  x1,
  y1,
  y2,
  x2 = x1,
  col = c("#C54E6D", "#009380"),
  lwd = c(1, 1),
  colgrid = NULL,
  mar = c(3, 6, 3, 6) + 0.1,
  ylab1 = paste(substitute(y1), collapse = ""),
  ylab2 = paste(substitute(y2), collapse = ""),
  nxbreaks = 5,
  yleg1 = paste(gsub("\\n$", "", ylab1), "(left axis)"),
  yleg2 = paste(ylab2, "(right axis)"),
  ylim1 = NULL,
  ylim2 = NULL,
  ylim.ref = NULL,
  xlab = "",
  main = NULL,
  legx = "topleft",
  legy = NULL,
  silent = FALSE,
  bty = "n",
  ...
)
}
\arguments{
\item{x1}{coordenadas da primeira linha ( data )}

\item{y1}{coordenadas da primeira linha ( data )}

\item{y2}{coordenadas da segunda linha ( dados )}

\item{x2}{coordenadas da segunda linha ( dados )}

\item{col}{cor padrao c("#C54E6D", "#009380")}

\item{lwd}{parametro do grafico (veja ?par)}

\item{colgrid}{xxxxxx}

\item{mar}{parametro do grafido (veja ?par)}

\item{ylab1}{rotulo entrada 1}

\item{ylab2}{rotulo entrada2}

\item{nxbreaks}{numero de quebras no eixo horizontal}

\item{yleg1}{rotulo do eixo}

\item{yleg2}{rotulo do eixo}

\item{ylim1}{tho numbers}

\item{ylim2}{tho numbers}

\item{ylim.ref}{yyyyyy}

\item{xlab}{e main são para rotulo x e titulo principal como em plot ()}

\item{main}{tho numbers}

\item{legx}{tho numbers}

\item{legy}{tho numbers}

\item{silent}{silent}

\item{bty}{bty}
}
\description{
DUAL PLOT FUNCTION
}
