% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dados.R
\name{getDados}
\alias{getDados}
\title{FUNCAOO PHDRISK DE SINALIZACAO DE NEGOCIOS}
\usage{
getDados(carteira, from, to, auto = FALSE)
}
\arguments{
\item{carteira}{base de dados}

\item{from}{data inicial}

\item{to}{data final}

\item{auto}{auto}
}
\description{
GET SYMBOL
}
