% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trad_especial.R
\name{tradeEspecial}
\alias{tradeEspecial}
\title{FUNCAO PHDRISK DE SINALIZACAO DE NEGOCIOS}
\usage{
tradeEspecial(ativo, n, plotar = TRUE)
}
\arguments{
\item{ativo}{base de dados}

\item{n}{INDEFINIDO}

\item{plotar}{TRUE/FALSE}
}
\description{
TRADE SINAL ESPECIAL
}
