% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trad_sinal_cci.R
\name{tradeSinalCci}
\alias{tradeSinalCci}
\title{FUNCAOO PHDRISK DE SINALIZACAO DE NEGOCIOS}
\usage{
tradeSinalCci(ativo, n, plotar = TRUE)
}
\arguments{
\item{ativo}{base de dados}

\item{n}{INDEFINIDO}

\item{plotar}{TRUE/FALSE}
}
\description{
TRADE SINAL SAR
}
