% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trad_sinal_mm.R
\name{tradeSinalMedia}
\alias{tradeSinalMedia}
\title{TRADE SINAL}
\usage{
tradeSinalMedia(ativo, media1, media2, plotar = TRUE)
}
\arguments{
\item{ativo}{VETORES}

\item{media1}{VETORES}

\item{media2}{VETORES}

\item{plotar}{plotar/dados}
}
\description{
TRADE SINAL
}
\examples{
  
inicial <- Sys.Date() - 180
ultima <- Sys.Date()
ativo <- c("BOBR4.SA")
at<-quantmod::getSymbols(ativo, from=inicial, to=ultima, auto.assign=F)
head(at)
 
 
}
