% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trad_sinal_sar.R
\name{tradeSinalSar}
\alias{tradeSinalSar}
\title{FUNCAOO PHDRISK DE SINALIZACAO DE NEGOCIOS}
\usage{
tradeSinalSar(ativo, plotar = TRUE)
}
\arguments{
\item{ativo}{base de dados}

\item{plotar}{TRUE/FALSE}
}
\description{
TRADE SINAL SAR
}
