% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trad_sinal_wpr.R
\name{tradeSinalWpr}
\alias{tradeSinalWpr}
\title{FUNCAOO PHDRISK DE SINALIZACAO DE NEGOCIOS}
\usage{
tradeSinalWpr(ativo, n, plotar = TRUE)
}
\arguments{
\item{ativo}{base de dados}

\item{n}{INDEFINIDO}

\item{plotar}{TRUE/FALSE}
}
\description{
TRADE SINAL WPR
}
